backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "gunicorn>=23.0.0",
    "requests>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "supabase>=2.0.0",
]
